+++ test/CommandLine.Test/InstallCommandTest.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Net;
 using Moq;
 using NuGet.Commands;
 using NuGet.Common;
@@ -601,8 +600,8 @@ namespace NuGet.Test.NuGetCommandLine.Commands
             var package1 = PackageUtility.CreatePackage("Foo", "1.0.0");
             var package2 = PackageUtility.CreatePackage("Qux", "2.3.56-beta");
             var repository = new MockPackageRepository { package1, package2 };
-            packageManager.Setup(p => p.InstallPackage(package1, true, true)).Verifiable();
-            packageManager.Setup(p => p.InstallPackage(package2, true, true)).Verifiable();
+            packageManager.Setup(p => p.InstallPackage(package1, true, true, true)).Verifiable();
+            packageManager.Setup(p => p.InstallPackage(package2, true, true, true)).Verifiable();
             packageManager.SetupGet(p => p.PathResolver).Returns(pathResolver);
             packageManager.SetupGet(p => p.LocalRepository).Returns(new LocalPackageRepository(pathResolver, fileSystem));
             packageManager.SetupGet(p => p.FileSystem).Returns(fileSystem);
@@ -635,7 +634,7 @@ namespace NuGet.Test.NuGetCommandLine.Commands
             var pathResolver = new DefaultPackagePathResolver(fileSystem);
             var packageManager = new Mock<IPackageManager>(MockBehavior.Strict);
             var repository = new MockPackageRepository { package };
-            packageManager.Setup(p => p.InstallPackage(package, true, true)).Verifiable();
+            packageManager.Setup(p => p.InstallPackage(package, true, true, true)).Verifiable();
             packageManager.SetupGet(p => p.PathResolver).Returns(pathResolver);
             packageManager.SetupGet(p => p.LocalRepository).Returns(new LocalPackageRepository(pathResolver, fileSystem));
             packageManager.SetupGet(p => p.FileSystem).Returns(fileSystem);
@@ -673,7 +672,7 @@ namespace NuGet.Test.NuGetCommandLine.Commands
             packageManager.SetupGet(p => p.LocalRepository).Returns(new LocalPackageRepository(pathResolver, fileSystem));
             packageManager.SetupGet(p => p.FileSystem).Returns(fileSystem);
             packageManager.SetupGet(p => p.SourceRepository).Returns(repository);
-            packageManager.Setup(p => p.InstallPackage(package, true, true)).Verifiable();
+            packageManager.Setup(p => p.InstallPackage(package, true, true, true)).Verifiable();
             var repositoryFactory = new Mock<IPackageRepositoryFactory>();
             repositoryFactory.Setup(r => r.CreateRepository("My Source")).Returns(repository);
             var packageSourceProvider = Mock.Of<IPackageSourceProvider>();
@@ -752,7 +751,7 @@ namespace NuGet.Test.NuGetCommandLine.Commands
             // We *shouldn't* be testing if a sequence of operations worked rather that the outcome that satellite package was installed correctly, 
             // but doing so requires work with  nice to have a unit test that tests it. 
             bool langPackInstalled = false;
-            packageManager.Setup(p => p.InstallPackage(package1, true, true)).Callback(() =>
+            packageManager.Setup(p => p.InstallPackage(package1, true, true, true)).Callback(() =>
                 {
                     if (langPackInstalled)
                     {